Require that names and namespaces are one global space across all semantic convention areas (#832)
* Require that names and namespaces are one global space across all semantic convention areas

We have semantic conventions for Resources, Spans and Metrics (in the future also Logs are expected).
It was not clear if the attribute names across all convention areas should be globally unique.

This commit asserts that conventions are one space, they are not independent spaces with
their own namespaces each.

We prohibit using the same Span or Resource attribute name or metric label name but give them
slightly different meanings or value sets.

Resolves: https://github.com/open-telemetry/opentelemetry-specification/issues/815

* Address PR comments

Co-authored-by: Bogdan Drutu <bogdandrutu@gmail.com>