feat: Add MySQL instance address attribute
At GitHub we run MySQL in multi-node clusters and are often interested in
knowing the hostname of a node that is executing a query or command.

In order to acheive this we added an attribute to the Trilogy instrumentation
named `db.mysql.instance.address` which is populated with the hostname of the node:

<https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/487>

This change adds the attribute to the database model as a MySQL specific connection attribute.