Script to check all specification links use the same version.
As described in https://github.com/open-telemetry/semantic-conventions/issues/46 this script checks all links to the specification follow a format where we know they are pointing at the same version as the rest of this repository. If the script is doing what is expected we can iterate to make it provide more specific information on the findings.