Fix problem in `xray-lambda` propagator definition
the problem comes from this "late" addition:
> To avoid potential issues when extracting with an active span context, the xray-lambda propagator SHOULD check if the provided context already has an active span context. If found, the propagator SHOULD return the provided context unmodified.

This is a problem because `xray-lambda` can't check the original context, only the context that is provided to it, which is modified by prior propagators in the chain.

This means if the original carrier has an x-ray context and an x-ray environment variable set, and if the xray propagator is configured as we expect it to be, then the xray-lambda will be either overwritten (if configured first), or will no-op because it detects an existing span context.

Proposed solution: change `xray-lambda` to replace `xray` in the configured list instead of work together with. This way, it is able to check the context before `xray` modifies it.