feat!: Update GraphQL span name convention
Due to concerns about high cardinality with the current span name format
use only the <graphql.operation.type> attribute for the span name.
The higher cardinality attribute, <graphql.operation.name> can still be
used in queries, as it is an attribute on the span.
The original fallback, "GraphQL Operation" is preserved.