System Semantic Conventions Non-Normative Guidance
This PR adds non-normative guidance from the System Semantic Conventions
Working Group. This is added in a new `groups` folder in
`non-normative`, and a `system` subfolder in `groups`. The docs written
here were already discussed in a Google doc where we were originally
collaborating on this, a link to which can be shared directly if
needed.