[chore] Remove duplicate `check` target
This PR removes a `check` target that I added without realizing that
there already was a target called `check`. The original intention was to
add a target that could be run locally without `markdown-link-check` as
that used to take a considerably long time. Since that is a fair bit
quicker now, there isn't much reason for it anymore. This PR essentially
undoes what I did when I added the check.