[chore] System Semconv Description Guidance
This PR explains how the `brief` and `note` fields will be used in the
namespaces covered by the System Semantic Conventions group.

The general concept is that we want to assume a baseline level of
knowledge for readers of our conventions, as this frees us up to focus
the descriptions of our instrumentation on instrumentation details
rather than on explaining general concepts.